[ADD] resource_advanced: new calendar type (full day) and unavailabilities
This commit introduces 2 new concepts in the resource time handling:

1/ Full day of Work, in calendar
Purpose is to avoid to lost one second per day (the last one) when working
several days in a row. So, to configure the attendance on a calendar, the
user can select directly entire days, without setting time slots.
This will help rental app.

2/ Get resource's unavailabilities
Get can now get the unavailabilies of a resource, this means the time
where the resource is not available to work (not attendance or leaves).
Purpose is to display them in grey on some views (like gantt or calendar)
to highlight those periods to end user.