[ADD] rental,sale_rental: book products and rent them
This commit introduces 2 new modules to handle the rental in odoo, without
being stock dependant (services only).

- rental: handle the rental booking and the resource management and
calendar. The lifecycle of a booking is draft, confirmed, picked up,
returned, done and cancel. The booking is simply the usage of the resource:
where and when it is rented.

- sale_rental: sell the rent time of resources. Resources can be attached
to product, and sell. the unit price depends on rental tenure defined on
product template, and depends on the rental period.

Limitation:
- rental can overlap. No hard constraint, just warning.
- pricelist are not yet supported, but all column and hooks are existing.
- product variant is not working 100% correclty.
- rental for none tracked consumable (no trace in the system of when to
pick up/return them).