[IMP] cloud_base: monkey patch db_filter
To ease database identification based on a request, we will allow nginx to
add custom header in request with the database name. The commit mokey patches
'db_filter' method in order to force the database to use.
Note: the database selector is not activated on production.

This is inspired from OCA module
(https://github.com/OCA/server-tools/blob/12.0/dbfilter_from_header/override.py)