[ADD] event_project,event_project_template: event generate project
Check the 'create project' option on event template, and the event will generate an
project on event creation. This creation can be base on a project template. Also, the
user will have to set the policy to determine the project start and stop dates.

The purpose is to setup a project in order to organize the event preparation through
tasks.