[ADD] resource_advanced: full day working calendar
This commit introduces a new mode to define work attendances on a working
calendar. User can now set full day as working day, avoiding the 1 second
missing (non consecutive days with native mode).
This module also provide some utilities method and field for resource time,
and bring the concept of 'unavailabilities' for a resource.
We can now search if a resource is available for a given period. This search
method must still be extended to cover more calendar mode, in a perfomant
way.

Resource view are completely redefined as we now want to expose them to end
user for a business goal, not not just in debug mode (in settings).