[ADD] rental: rent your resource and manage their calendar
This module allows end user to rent its resource for a given period. This
means the resource is away from the company, at a customer's place.
The rental bookings are displayed on a gantt view, and have a chatter/activity
scheduling. This module only handle resource planning and booking lifecycle. A
booking is seen as a time entry in the resource working calendar.