[IMP] sale_rental: rework pricing, pricelist and discount in rental flow
This commit does a few things:
1/ clarify price computation: rental price must be compute in the pricelist
currency, by summing converted tenure to avoid rounding mistakes. No pricelist
must be applied during that. Tenure sum is the price unit. Then the pricelist
must be applied. A discount can be extract from the pricelist application, as
in the sale flow (see discount_policy of the pricelist).

2/ discount was not appearing of SO line form view: discount is either included
in unit price, either exposed in the discount column (in eCommence, prices are
marked in red).

3/ Rework the onchanges. This is such a spaghetti code. Also we used to set the
price unit in the compute method of another field. This is not correct. The
onchange flow are now completely split: sale vs rental.
It will ease the discount computation and set up when pricelist will be applied.

4/ This commit also prepares the coming of pricelist management for rental price,
even for the website use case. (Let's hope !)