[ADD] pos_restaurant_kitchen: new module
This commit provides a new module to organize pos order line in a kanban
view, for preparation purpose. The idea is to get the preparation status
of order parts in allowed locations (kitchen, bar, ...).
The printer are used as location now: so you can have a "location" that
is not printing stuff, but see the order parts of selected product
categories.

The kanban is refresh automatically, with a context key with the delay to
use.

This commit also provides a new user group "cooker" for the user to only
see the order parts on the kitchen kanban.