[IMP] web_view_gantt: rename to web_gantt
The purpose of this commit is to have a more similar gantt view to
the one of enterprise edition. This way we can easily exchange module
and view definition.
The type of the view is now 'gantt'.

However module from enterprise extending the gantt view (enterprise
implementation) will failed, as our implement still rely on DHXHTML
lib.