event_project: allow to create task in exising project
A project can be created when the event arrived in a special stage,
based or not on a project template. Now, we add the option to link
the event to an existing project. Tasks will be created manually,
from the stat button of event form view. The task are linked to the
event.