event_advanced: qrcode on badge
QRcode are an option on the badge. We can define it on the template,
and the event. Then, the QRCode appears on the badge.
Scanning the QRCode will mark the registration as attended, in a
separated web page exposing the actual registration status.
This page requires registration access right, and being logged in.