[REF] *rental: remove inheritS on rental model
This commit removes the inheritS on the rental booking model. So, the entry in
the resource calendar will be created on booking confirmation.
The state of rental book is also simplified: a draft booking can be confirmed,
then marke as done. It can be cancelled anytime, and reset to draft to be
removed.

Carreful: tests are not updated !!!