[IMP] document: introduces request for a document
Now that the inheritS is removed, request document (not linked to attachment) can
be introduced.
When creating a request, a next activity is generated (type is upload a
document). When the action is done, the file is set to the document and
the request change its type. After that, it is impossible to create
another next activity 'upload file' on the document (since it is not a
request anymore).