[IMP] *event*: sale quantity and pre order ticket
This commit brings 2 features in the event modules flow.

1/ quantity on registration: inspired from OCA module, we can now book
and sell 1 regsitration with a quantity on it. This aims to avoid the
user to introduced all contact details of an attendee for each
regristration. Added the website part (that was missing in the OCA
module).

2/ allow to pre-order ticket: this means that even if a price is display
on the website, no sales order will be generated (or those line will not
be included in the current cart). This implementation might not be
efficient as we need to destroy the SO created by the standart flow
defined in odoo 15 (no good hook to prevent SO creation, ....).